Use 'Encoding' for reading and writing, #53
When creating a `description` object, `desc` observes the `Encoding`
field, if present, and uses the specified encoding to parse the file.
Internally, it converts all fields to UTF-8.

When writing a `description` object to a file, `desc` uses the
`Encoding` field (if present), and converts all fields to the specified
encoding.