Fix text inputs with marked encodings
If `desc()` was called with a text parameter and the input was _not_ in
the default encoding for the locale `textConnection()` would be called
with the default encoding parameter of `""`, which attempts to convert
the encoding to the default one. Using encoding = "bytes" means the
connection does not try to convert the encoding and just passes the
bytes through unchanged, which fixes the issue.